requirements.txt
django==2.0.3
django-redis==4.9.0
djangorestframework==3.7.7
